# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart-Browser Bookmarks API",
    "description": "Tag-first bookmarks. Contract version 2. See doc/architecture.md.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/bookmarks": {
      "post": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Save Bookmark",
        "description": "Save a page.\n\n201 the first time, 200 every time after, never a second row \u2014 whatever the spelling\nof the URL and however many clients call at once. Tags in the payload are merged on\nevery call, so re-saving with new tags adds them rather than failing.",
        "operationId": "save_bookmark_api_v2_bookmarks_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "List Bookmarks",
        "description": "List your saves.\n\n`mode=all` is the intersection the tag sidebar is built on \u2014 `python` AND `fastapi`.\n`untagged=true` is how you work the backlog.",
        "operationId": "list_bookmarks_api_v2_bookmarks_get",
        "parameters": [
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated tag names",
              "title": "Tags"
            },
            "description": "Comma-separated tag names"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "any"
              ],
              "type": "string",
              "default": "all",
              "title": "Mode"
            }
          },
          {
            "name": "site",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on title or URL",
              "title": "Q"
            },
            "description": "Substring match on title or URL"
          },
          {
            "name": "untagged",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only bookmarks with no tags",
              "default": false,
              "title": "Untagged"
            },
            "description": "Only bookmarks with no tags"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/lookup": {
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Lookup Bookmark",
        "description": "Have I saved this page, and with what tags?\n\nA read. `POST /bookmarks` is idempotent and would answer the same question, but it is\na write: a client calling it merely to ask would create a bookmark every time someone\nopened a popup. 404 means not saved, which is a normal answer here.",
        "operationId": "lookup_bookmark_api_v2_bookmarks_lookup_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Url"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{save_id}": {
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Get Bookmark",
        "operationId": "get_bookmark_api_v2_bookmarks__save_id__get",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Patch Bookmark",
        "operationId": "patch_bookmark_api_v2_bookmarks__save_id__patch",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Delete Bookmark",
        "description": "Unsave a page.\n\nA soft delete of *your* save. The `bookmark` row and anything crawled for it survive,\nbecause they belong to everyone; and re-saving the page restores this row rather than\ncreating a second one.",
        "operationId": "delete_bookmark_api_v2_bookmarks__save_id__delete",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{save_id}/tags": {
      "post": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Add Tags",
        "operationId": "add_tags_api_v2_bookmarks__save_id__tags_post",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{save_id}/tags/{tag_name}": {
      "delete": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Remove Tag",
        "operationId": "remove_tag_api_v2_bookmarks__save_id__tags__tag_name__delete",
        "parameters": [
          {
            "name": "save_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Save Id"
            }
          },
          {
            "name": "tag_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookmarkCreate": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "saved_from": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved From"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "BookmarkCreate"
      },
      "BookmarkOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "saved_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved From"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "site",
          "saved_from",
          "created_at",
          "tags"
        ],
        "title": "BookmarkOut"
      },
      "BookmarkPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookmarkOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "BookmarkPage"
      },
      "BookmarkPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "BookmarkPatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TagsIn": {
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "user",
              "ai",
              "rule",
              "import"
            ],
            "title": "Source",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "TagsIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}